pyproject.toml
[tool.poetry]
name = "djangotestproject"
version = "0.1.0"
description = ""
authors = ["Aliaskar <alyas.toronov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
